Review fix: the watch is installed before anything can be reading it
Keeping the logger alive fixed teardown and left the other end of the same race.
instance() was first called from the corpus test, by which time earlier suites
have initialised the shared engine and its long-lived threads, so that first
call wrote juce::Logger's unsynchronised pointer while one of them could have
been reading it.

It moves to AssertionWatch.hpp and is installed from main(), before the JUCE
initialiser and before anything exists that could start a thread. Being first
and staying for ever is what removes the race; there is no lock on this side
that would, because the read happens inside JUCE.
